Implement CheckNameConflict for Unix
Make CheckNameConflict system-independent and move system-dependent code into function FindRealFileName.
